DE937 handle multi-command translations
Needed to translate 'ip netns list' command to multiple command
in Mercury environment, as we have some services in DHCP container.
To simplify this, I decided to allow translation to include multiple
commands using ';;;' as a separator. Each such command will be handled
separately. It's not efficient but it works and is much more simple
than trying to use "&&" in the translation.