US3139 carry forward constraint attributes on implicit links
When a link has a constraint attribute such as "network", it needs to persist in the implicit links created from it.
If the other link in the implicit link does not have the same attribute - the implicit link will have the same attribute with same value.
For example, if link of type "instance-vnic" is on network X, and the matching "vnic-vconnector" does not have "network", the "instance-vconnector" implicit link will have "network: X" attribute.
If the other link in the implicit link has the same attribute with the same value - the implicit link will have the same attribute with same value.
If the other link in the implicit link has the same attribute with another value - the implicit link will not be created.